spendwise_backend/requirements.txt
asgiref==3.11.0
Django==3.1.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djongo==1.3.7
pymongo==3.11.4
pytz==2025.2
setuptools==80.9.0
sqlparse==0.2.4
tzdata==2025.3
